Issue-600: Extract common default thread pool config, increase antivirus query to 50
